Bug 369863 - Add OAuth 1.0a support to the OAuth framework
The first version of the framework only supported OAuth 1.0
authentication flows. We should add 1.0a support (and add a warning to
the login page if consumers are using 1.0). The OAuth version can be
negotiated at runtime depending on whether the consumer passes an
oauth_callback parameter when asking for a request token.