Bug 379591 - Alternate approach for hostname and public URI handling.
There are three possibilities:

1.  org.eclipse.lyo.oslc4j.publicURI is set.   This will take
precedence.

2.  publicURI is not set and
org.eclipse.lyo.oslc4j.disableHostResolution
is either a) not set at all or b) set to false.   In this case we'll try
to resolve the local hostname with getCanonicalHostname

3.  org.eclipse.lyo.oslc4j.disableHostResolution is a) set to true or b)
hostname resolution failed in scenario 2.  In this case we fallback and
use the hostname off of the request.

Change-Id: Ia9b1764210e1982821a18af8fa23716e639f3f7a