Bug 36697 - StackOverflowException when extended property refers to the current resource
Break recursion in both the RDF/XML and JSON helpers when there are
cycles in the resource references in the model.  Cycles can be caused by
URI references in the model and by blank nodes.

This also fixes the issue where OSLC4J was creating multiple Java
objects for an in-bound blank node which was referenced multiple times
in the model.

Change-Id: I1eaaa0d0da863a7bb8e4b27282385a9c745decec