Bug392951  AutomationAdapter sample should reuse existing registration when possible
When an AutomationAdapter registers with an Automation Service Provider
a server-side resource is created for it.  The URL for this server-side
resource is reflected in the AutomationAdapter's "about" property.  If
the AutomationAdapter is started again then it should reuse the existing
server side resource instead of creating a new one.  The URL can be
persisted with the other adapter properties such as the project name and
adapter credentials.