Bug 396890 - Allow finer grained registration of Producers and Consumers
Developers may have the need to not support all media types OSLC4J
supports or to replace OSLC4J Producers and Consumers with their own. 
Create sub-classes of the OslcRdfXml* classes to separate XML and Turtle
serializations.   Ideally we'd rename/re-factor the OslcRdfXml* classes
as well, but that is not part of this commit.


Change-Id: Id084a607284ef1949dc7f0e4ded44b45c7839ab5
Signed-off-by: Michael Fiedler <fiedler.mf@gmail.com>