Add launch configuration for building OSLC4J
m2eclipse uses the Eclipse JRE to build. This causes problems in some
environments where the javadoc command can't be found, notably OS X,
even when JAVA_HOME is set in the shell. To workaround, provide a launch
for building OSLC4J that skips Javadoc generation inside of Eclipse.
This is really only important for releases anyway.

Change-Id: I0e73910bb50d87628e86c1e83c8c01ae141466c0
Signed-off-by: Samuel Padgett <spadgett@us.ibm.com>