Bug 466402 - Generator: Request: Make it possible to generate backend code for getters/setters
Added 4 new attributes to the AdaptorInterface Eclass:
* backendCodeTemplate_getResources
* backendCodeTemplate_getResource
* backendCodeTemplate_searchResources
* backendCodeTemplate_createResource

These should allow the end-user to specify a pattern of code that can
get/search/create a resource from the backend tool being interfaced.
This is meant to replace (some of) the code that is normally manually
implemented in the AdaptorManager class.

The template currently takes 2 keywords [ResourceClassName/] &
[ResourceInstanceName/].

Also, the allRelatedResources(aResource: Resource) query has been
modified. Fixing an old odd bug related to Bug466401 - which somehow
ended up in this branch.

Change-Id: Ife8b1f5451b21622bcdf0bc91c4c9c0258b3944a
Signed-off-by: Jad El-khoury <jad@kth.se>