Bug 489703 - Generator: support the definition and generation of client code
Basic support for the generation of client code implemented.
To define a client, the modeller can either
(a) reference another AdaptorInterface model
(a) define a client as a set of resources for which access functions
need to be defined.

Currently, only the basic functions to GET the ServiceProviderCatalog
and each serviced resources are generated.

Change-Id: I7e0427949d03331a7d126421b2bda86c5fb2a9af
Signed-off-by: Jad El-khoury <jad@kth.se>