Bug 513477 - Replace log4j with slf4j in core
And do some small refactoring by extracting code to methods as well as
inverting if conditions (eg if(!a){then}else{return} becomes if(a){return}then).
This saves us space in case of many nested blocks and eliminates the
need to keep all the if conditions in the memory while reading the code.

Exclude slf4j-log4j12 (concrete impl) from the used Jena libraries (they
have fixed the issue only in Jena 3).

Change-Id: I1e4a74074647a1f8f929bbc0691c20fecc921dfe
Signed-off-by: Andrew Berezovskyi <andriib@kth.se>