Added support for JEE Forms base Authentication
This was done by adding OslcClient.login(userId, password) which is called by the client applications after the rootservices document has been read (which doesn't require authentication).

I will still see if there's a way to get this to work with a registered authenticator filter, but at least its working, and maybe explicit login is ok, or perhaps preferred because it avoids having a filter run on every request.