JEEFormAuthenticator is now working and tested with all jazz-app clients
The JEEFormAuthenticator provides a filter that detects JEE Form based authentication challenges and does the user authentication when needed.

OslcClient.login() has been removed so that there is no authentication dependencies in OslcClient. All authentication is handled by the client applications through authentication filters registered with the ClientBuilder.

All the jazz-app samples have been updated to use the new OslcClient. Each of them provides an example of how to configure HttpClient (instead of HttpUrlConnection), SSL to trust self-assigned certificates, and use of the JEEFormAuthenticator for authorization.

The launch configurations have been updated and all the applications run properly.