Generic TypeVariable handling (#286)
When we want to return a Response instead of a `Collection<AbstractResource>`, we need to use `ResponseBuilder.entity()` method to pass the entity. The generic signature is not preserved this way. The suggested way to solve this is to wrap the entity with a `GenericEntity`.

Unfortunately, this changes the type of the generic arguments as seen by JAX-RS Providers from `Class` to `TypeVariable`. This PR enables RDF-based providers to handle both kinds. It also adds error log statements for the JSON Provider that the `TypeVariable` support is missing.